Refactor: Separate Linear and CLI adapters into subdirectories
Reorganized packages/core/src/issue-tracker/adapters/ to improve
maintainability and code organization:

- Created adapters/linear/ directory for Linear-specific implementations
- Created adapters/cli/ directory for CLI platform implementations
- Moved LinearIssueTrackerService, LinearTypeAdapters, and
  LinearAgentEventTransport to linear/ subdirectory
- Moved CLIIssueTrackerService, CLIEventTransport, and CLIRPCServer
  to cli/ subdirectory
- Moved all corresponding test files to their respective __tests__/
  subdirectories
- Updated all import paths throughout the codebase to reflect new
  structure
- Updated export paths in issue-tracker/index.ts

All 58 tests passing, TypeScript compilation clean.

Fixes CYPACK-326
